+++ R/rate_limit.R
@@ -8,6 +8,9 @@
 #' @examples
 #' ct_get_remaining_hourly_queries()
 ct_get_remaining_hourly_queries <- function() {
+  # If the hourly time limit has past, reset the hourly time limt and hourly
+  # query count.
+  ct_get_reset_time()
   get("queries_this_hour", envir = ct_env)
 }
 
@@ -15,6 +18,10 @@ ct_get_remaining_hourly_queries <- function() {
 #' Comtradr rate limit time check
 #'
 #' Get the time in which the hourly limit will reset.
+#'
+#' @param set logical, if \code{TRUE} and the current reset time is
+#'   \code{NULL}, set the reset time to be one hour from the current
+#'   \code{Sys.time}.
 #'
 #' @return date and time in which the hourly query limit will reset. Return is
 #'   a "POSIXct" object (see \code{\link{DateTimeClasses}}).
@@ -25,13 +32,58 @@ ct_get_remaining_hourly_queries <- function() {
 #'
 #' # Get minutes remaining until limit reset, as numeric value.
 #' as.double(ct_get_reset_time() - Sys.time())
-ct_get_reset_time <- function() {
+ct_get_reset_time <- function(set = NULL) {
+  stopifnot(is.null(set) || is.logical(set))
   val <- get("next_hour_reset", envir = ct_env)
-  if (!is.null(val)) {
-    val + 3600
-  } else {
-    Sys.time() + 3600
+  if (is.null(val)) {
+    if (isTRUE(set)) {
+      # If val is NULL and set is TRUE, reset the time limit in the pkg env,
+      # then return new value.
+      reset_hourly_limits()
+      return(get("next_hour_reset", envir = ct_env))
+    }
+    # If val is NULL and set is not TRUE, return current time plus one hour.
+    return(Sys.time() + 3600)
   }
+  if (Sys.time() > val) {
+    # If "next_hour_reset" time limit has expired, reset the
+    # time limit in the pkg env, then return new value.
+    reset_hourly_limits()
+    return(get("next_hour_reset", envir = ct_env))
+  }
+
+  # Else if the time limit has not yet expired, return the current value.
+  return(val)
+}
+
+
+#' Reset Hourly Limits
+#'
+#' Reset the "next_hour_reset" and "queries_this_hour" values in the pkg env.
+#'
+#' @noRd
+reset_hourly_limits <- function() {
+  assign("next_hour_reset", Sys.time() + 3600, envir = ct_env)
+  assign("queries_this_hour", getOption("comtradr")$comtrade$per_hour_limit,
+         envir = ct_env)
+}
+
+
+#' Validate Token
+#'
+#' Verifies validity of token provided by user.
+#'
+#' @param token char string, API token
+#'
+#' @return Returns token if successfully validated by API; NULL otherwise.
+#'
+#' @noRd
+validate_token <- function (token) {
+  result <- NULL
+  tryCatch(result <- jsonlite::fromJSON(
+    paste0("https://comtrade.un.org/api/getUserInfo?token=", token)),
+    error = function(e) e)
+  if (!is.null(result)) token else result
 }
 
 
@@ -52,6 +104,9 @@ ct_get_reset_time <- function() {
 ct_register_token <- function(token) {
   # input validation.
   stopifnot(is.character(token) || is.null(token))
+
+  # Verify validity of token
+  token <- validate_token(token)
 
   # Get number of queries left for the current hour.
   queries_this_hour <- get("queries_this_hour", envir = ct_env)
@@ -77,6 +132,9 @@ ct_register_token <- function(token) {
     return(invisible())
   }
 
+  # Get the current account type.
+  curr_account_type <- getOption("comtradr")$comtrade$account_type
+
   # Set token within options.
   ct_options <- getOption("comtradr")
   ct_options$comtrade$token <- token
@@ -87,11 +145,10 @@ ct_register_token <- function(token) {
 
   # Change the hourly limit within the env ct_env. Subtract the number of
   # queries already performed this hour from the update value.
-  curr_account_type <- getOption("comtradr")$comtrade$account_type
   if (curr_account_type == "standard") {
     new_hr_limit <- 10000 - (100 - queries_this_hour)
   } else if (curr_account_type == "premium") {
-    new_hr_limit <- 10000 - (10000 - queries_this_hour)
+    new_hr_limit <- queries_this_hour
   }
   assign("queries_this_hour", new_hr_limit, envir = ct_env)