+++ inst/doc/comtradr-vignette.R
@@ -1,38 +1,38 @@
-## ---- echo = FALSE-------------------------------------------------------
+## ---- echo = FALSE------------------------------------------------------------
 knitr::opts_chunk$set(comment = "#>", collapse = TRUE, fig.width = 9, fig.height = 6)
 
-## ----eval = FALSE--------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  install.packages("comtradr")
 
-## ----eval = FALSE--------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  devtools::install_github("ChrisMuir/comtradr")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 library(comtradr)
 
-## ---- echo = FALSE-------------------------------------------------------
+## ---- echo = FALSE------------------------------------------------------------
 v_data_1 <- system.file("extdata", "vignette_data_1.rda", package = "comtradr")
 if (!file.exists(v_data_1)) {
   stop("internal vignette data set '~/extdata/vignette_data_1.rda' not found", call. = FALSE)
 }
 load(v_data_1)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  q <- ct_search(reporters = "USA",
 #                 partners = c("Germany", "France", "Japan", "Mexico"),
 #                 trade_direction = "imports")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 str(q)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  q <- ct_search(reporters = "USA",
 #                 partners = c("Germany", "France", "Japan", "Mexico"),
 #                 trade_direction = "imports",
 #                 start_date = 2010,
 #                 end_date = 2014)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  # Get all monthly data for a single year (API max of 12 months per call).
 #  q <- ct_search(reporters = "USA",
 #                 partners = c("Germany", "France", "Japan", "Mexico"),
@@ -49,19 +49,19 @@ str(q)
 #                 end_date = "2012-07",
 #                 freq = "monthly")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_country_lookup("korea", "reporter")
 ct_country_lookup("bolivia", "partner")
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  q <- ct_search(reporters = "Rep. of Korea",
 #                 partners = "Bolivia (Plurinational State of)",
 #                 trade_direction = "all")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_commodity_lookup("tomato")
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  tomato_codes <- ct_commodity_lookup("tomato",
 #                                      return_code = TRUE,
 #                                      return_char = TRUE)
@@ -71,13 +71,13 @@ ct_commodity_lookup("tomato")
 #                 trade_direction = "all",
 #                 commod_codes = tomato_codes)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  q <- ct_search(reporters = "USA",
 #                 partners = c("Germany", "France", "Mexico"),
 #                 trade_direction = "all",
 #                 commod_codes = c("0702", "070200", "2002", "200210", "200290"))
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 # The url of the API call.
 attributes(q)$url
 # The date-time of the API call.
@@ -86,49 +86,49 @@ attributes(q)$time_stamp
 # The total duration of the API call, in seconds.
 attributes(q)$req_duration
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_country_lookup(c("Belgium", "vietnam", "brazil"), "reporter")
 
 ct_commodity_lookup(c("tomato", "trout"), return_char = TRUE)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_commodity_lookup(c("tomato", "trout"), return_char = FALSE)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_commodity_lookup(c("tomato", "sldfkjkfdsklsd"), verbose = TRUE)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  ct_update_databases()
 #  #> All DB's are up to date, no action required
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ct_commodity_db_type()
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 # Column headers returned from function ct_search
 colnames(q)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 # Apply polished column headers
 q <- ct_use_pretty_cols(q)
 
 # Print new column headers.
 colnames(q)
 
-## ---- echo = FALSE-------------------------------------------------------
+## ---- echo = FALSE------------------------------------------------------------
 v_data_2 <- system.file("extdata", "vignette_data_2.rda", package = "comtradr")
 if (!file.exists(v_data_2)) {
   stop("internal vignette data set '~/extdata/vignette_data_2.rda' not found", call. = FALSE)
 }
 load(v_data_2)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  # Comtrade api query.
 #  df <- ct_search(reporters = "China",
 #                  partners = c("Rep. of Korea", "USA", "Mexico"),
 #                  trade_direction = "exports")
 
-## ---- warning = FALSE, message = FALSE-----------------------------------
+## ---- warning = FALSE, message = FALSE----------------------------------------
 library(ggplot2)
 
 # Apply polished col headers.
@@ -147,14 +147,14 @@ ggplot(df, aes(Year, `Trade Value usd`, color = factor(`Partner Country`),
   labs(title = "Total Value (USD) of Chinese Exports, by Year") +
   theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1))
 
-## ---- echo = FALSE-------------------------------------------------------
+## ---- echo = FALSE------------------------------------------------------------
 v_data_3 <- system.file("extdata", "vignette_data_3.rda", package = "comtradr")
 if (!file.exists(v_data_3)) {
   stop("internal vignette data set '~/extdata/vignette_data_3.rda' not found", call. = FALSE)
 }
 load(v_data_3)
 
-## ---- eval = FALSE-------------------------------------------------------
+## ---- eval = FALSE------------------------------------------------------------
 #  # First, collect commodity codes related to shrimp.
 #  shrimp_codes <- ct_commodity_lookup("shrimp",
 #                                      return_code = TRUE,
@@ -168,7 +168,7 @@ load(v_data_3)
 #                  end_date = 2011,
 #                  commod_codes = shrimp_codes)
 
-## ---- warning = FALSE, message = FALSE-----------------------------------
+## ---- warning = FALSE, message = FALSE----------------------------------------
 library(ggplot2)
 library(dplyr)