+++ NEWS.md
@@ -1,3 +1,30 @@
+comtradr 0.3.0
+====================
+
+## NEW FEATURES
+
+* Modifications to `ct_search()` to add support for commodity code `ag6` ([#30](https://github.com/ropensci/comtradr/pull/30))
+
+## BUG FIXES
+
+* Function `ct_register_token()` now checks if the provided token is recognized by the official API and only grants "premium" credentials if it is ([#34](https://github.com/ropensci/comtradr/issues/34)).
+
+* Passing an API token string to `ct_register_token()` now properly bumps the hourly rate limit up to 10,000
+([#21](https://github.com/ropensci/comtradr/issues/21)).
+
+* In func `ct_search()`, passing a character vector of long-form commodity descriptions to arg `commod_codes` will now 
+throw an error prior to making an API call, which would fail ([#24](https://github.com/ropensci/comtradr/issues/24)).
+
+* Update the country package data, to stay up to date with the reporter/partner country table that Comtrade is using. This is an update to
+the file `inst/extdata/country_table.rda`. ([#29](https://github.com/ropensci/comtradr/issues/29)).
+
+* In func `ct_search()`, improve error messaging when an input country is invalid. ([#31](https://github.com/ropensci/comtradr/issues/31)).
+
+* In func `ct_search()`, fix bug in which running queries using the `SITCrev2` commodity type was returning raw HTML (as opposed to json data). ([#27](https://github.com/ropensci/comtradr/issues/27)).
+
+* In func `ct_country_lookup()`, remove `NA` inputs from multi-country lookups (previously `NA` was being stringified and included in the pipe-separated regex lookup).
+
+
 comtradr 0.2.2
 ====================
 
@@ -42,7 +69,7 @@ comtradr 0.1.0
 * Added package level man page.
 * Now using native R errors/warnings, as opposed to nesting API status codes in a returned list.
 * `Imports` changes: remove `dplyr`, add `magrittr` and `purrr`.
-* Expand and improve test coverage via [testthat](https://github.com/hadley/testthat).
+* Expand and improve test coverage via [testthat](https://github.com/r-lib/testthat).
 
 ## BUG FIXES
 
@@ -58,7 +85,7 @@ comtradr 0.0.2 (2017-07-03)
 
 ## MINOR IMPROVEMENTS
 
-* Add unit tests via [testthat](https://github.com/hadley/testthat).
+* Add unit tests via [testthat](https://github.com/r-lib/testthat).
 
 
 comtradr 0.0.1 (2017-04-06)